ml/training/requirements.txt
# Training package requirements

# Core ML dependencies
tensorflow>=2.13.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Image processing
Pillow>=9.0.0
scipy>=1.10.0

# Data handling
PyYAML>=6.0
pandas>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
tqdm>=4.65.0
